requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
faiss-cpu
streamlit
langchain_groq
langchain_chroma
chromadb
pypdf
langchain-huggingface